a lot

+++ src/main/java/ro/upt/ac/portofolii/portofoliu/PortofoliuController.java
@@ -1,14 +1,15 @@
 package ro.upt.ac.portofolii.portofoliu;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.*;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 import ro.upt.ac.portofolii.cadruDidactic.CadruDidactic;
 import ro.upt.ac.portofolii.cadruDidactic.CadruDidacticRepository;
 import ro.upt.ac.portofolii.student.Student;
@@ -16,6 +17,11 @@ import ro.upt.ac.portofolii.student.StudentRepository;
 import ro.upt.ac.portofolii.tutore.Tutore;
 import ro.upt.ac.portofolii.tutore.TutoreRepository;
 import ro.upt.ac.portofolii.tutore.TutoreService;
+import ro.upt.ac.portofolii.utils.PdfGenerator;
+import ro.upt.ac.portofolii.utils.WordGenerator;
+
+import java.util.ArrayList;
+import java.util.List;
 
 @Controller
 public class PortofoliuController
@@ -105,7 +111,27 @@ public class PortofoliuController
 	    model.addAttribute("portofolii", portofoliuRepository.findAll());
 	    return "portofoliu-read";
 	}
-	
+
+	@GetMapping("/student-portofoliu-read/{id}")
+	public String readPortofoliuStudent(@PathVariable("id") int id, Model model)
+	{
+		Student student = studentRepository.findById(id);
+		if(student == null)
+		{
+			throw new RuntimeException("Student ID not found");
+		}
+		List<Portofoliu> portfolios = portofoliuRepository.findAll();
+		List<Portofoliu> portfolioStudents = new ArrayList<>();
+		for(Portofoliu p : portfolios){
+			if(p.getStudent().getId() == student.getId()){
+				portfolioStudents.add(p);
+			}
+		}
+		model.addAttribute("student", student);
+		model.addAttribute("portofolii", portfolioStudents);
+		return "student-portofolii-read";
+	}
+
 	@GetMapping("/portofoliu-edit/{id}")
 	public String edit(@PathVariable("id") int id, Model model) 
 	{
@@ -130,11 +156,68 @@ public class PortofoliuController
 	}
 	
 	@GetMapping("/portofoliu-delete/{id}")
-	public String delete(@PathVariable("id") int id)
+	public String delete(@PathVariable("id") int id, RedirectAttributes redirectAttributes)
 	{
 	    Portofoliu portofoliu = portofoliuRepository.findById(id);
-	    
+		if (portofoliu == null) {
+			redirectAttributes.addFlashAttribute("errorMessage", "Portofoliul nu a fost găsit.");
+			return "redirect:/portofoliu-read";
+		}
+		portofoliu.setStudent(null);
+		portofoliu.setTutore(null);
+		portofoliu.setCadruDidactic(null);
 	    portofoliuRepository.delete(portofoliu);
 	    return "redirect:/portofoliu-read";
-	}	
+	}
+
+	@GetMapping("/portofoliu-view/{id}")
+	public String viewPortofoliu(@PathVariable("id") int id, Model model) {
+		Portofoliu portofoliu = portofoliuRepository.findById(id);
+
+		if (portofoliu == null) {
+			model.addAttribute("errorMessage", "Portofoliul nu a fost găsit.");
+			return "redirect:/portofoliu-read";
+		}
+
+		model.addAttribute("portofoliuId", id);
+		return "portofoliu-vizualizare";
+	}
+
+	@GetMapping("/portofoliu-generate-docx/{id}")
+	@ResponseBody
+	public ResponseEntity<byte[]> generatePortofoliuDocx(@PathVariable("id") int id) {
+		Portofoliu portofoliu = portofoliuRepository.findById(id);
+
+		if (portofoliu == null) {
+			return ResponseEntity.notFound().build();
+		}
+
+		byte[] pdfBytes = WordGenerator.generatePortofoliuDocx(portofoliu);
+
+        HttpHeaders headers = new HttpHeaders();
+		headers.setContentType(MediaType.APPLICATION_PDF);
+
+		return ResponseEntity.ok()
+				.headers(headers)
+				.body(pdfBytes);
+	}
+
+	@GetMapping("/portofoliu-generate-pdf/{id}")
+	@ResponseBody
+	public ResponseEntity<byte[]> generatePortofoliuPDF(@PathVariable("id") int id) {
+		Portofoliu portofoliu = portofoliuRepository.findById(id);
+
+		if (portofoliu == null) {
+			return ResponseEntity.notFound().build();
+		}
+
+		byte[] pdfBytes = PdfGenerator.generatePortofoliuPdf(portofoliu);
+
+		HttpHeaders headers = new HttpHeaders();
+		headers.setContentType(MediaType.APPLICATION_PDF);
+
+		return ResponseEntity.ok()
+				.headers(headers)
+				.body(pdfBytes);
+	}
 }
